api-backend/requirements.txt
asgiref==3.5.2
Django==4.1.3
djangorestframework==3.14.0
flake8==5.0.4
Markdown==3.4.1
mccabe==0.7.0
Pillow==9.3.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2022.6
sqlparse==0.4.3
